backend/requirements.txt
asgiref==3.7.2
autopep8==2.0.2
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.54.0
django-cors-headers==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.2
djangorestframework==3.14.0
Faker==19.3.1
flake8==6.0.0
gunicorn==21.2.0
h11==0.14.0
idna==3.4
Markdown==3.4.4
mccabe==0.7.0
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
urllib3==2.0.4
uvicorn==0.23.2
